Misc code style improvements and warning fixes
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@14819 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/graphics/callbacks.hpp
@@ -544,10 +544,10 @@ public:
         // Norm factor
         double w = rand();
         w /= RAND_MAX;
-        array[4 * i] = x;
-        array[4 * i + 1] = y;
-        array[4 * i + 2] = z;
-        array[4 * i + 3] = w;
+        array[4 * i] = (float)x;
+        array[4 * i + 1] = (float)y;
+        array[4 * i + 2] = (float)z;
+        array[4 * i + 3] = (float)w;
       }
     }
